refactor: 글로벌 스타일 지정 및 타입 선언 위치 변경
