+++ components/src/global.styled.ts
@@ -1,0 +1,12 @@
+import { createGlobalStyle } from 'styled-components';
+import reset from 'styled-reset';
+
+const GlobalStyle = createGlobalStyle`
+  ${reset}
+  
+  body {
+    line-height: 1.2;
+  }
+`;
+
+export default GlobalStyle;
